Delete the target instead of replacing it
In some circumstances, we may want to replace a running subctl,
e.g. if the running "subctl" is really a download script (see
https://github.com/submariner-io/shipyard/pull/893 for the scenario
which prompted this).

cp -f only deletes the target if it can't be written. Instead, remove
the target explicitly, using rm -f so we don't fail if it doesn't
exist. This ensures that the old inode is preserved if it is still in
use, and the new file doesn't replace an old one which is still
needed.

On platforms with GNU cp, cp --remove-destination could be used
instead, but getsubctl.sh should work on non-GNU platforms too.

Signed-off-by: Stephen Kitt <skitt@redhat.com>