Handle .gz downloads
Now that the release process publishes gzip-compressed tarballs
alongside the xz-compressed ones, getsubctl needs to handle both.

When decompressing, if a download matches .tar.gz, pass -z to tar to
decompress it (since the decompression consumes piped data, tar needs
to be told about the format ahead of time, so this can't use a generic
.tar.* handler relying on tar's auto-detection).

When downloading, prefer .xz tarballs.

Signed-off-by: Stephen Kitt <skitt@redhat.com>